refactor: split test file + template polish + MCPTool godoc clarity
Second /simplify pass — addresses findings after the mirrorstack.go split.

Test split (symmetry with production files):

  mirrorstack_test.go   1052 → 709 lines — retains core/scope/lifecycle
                        tests (New, Platform/Public/Internal, Manifest,
                        RequirePermission, Init, MaxBytes, ScopesPanic,
                        ModuleSchema).
  describe_test.go       NEW — 221 lines — Describe/DependsOn/Needs/
                         Resolve/parseDepSpec tests including SemVer
                         constraint coverage.
  mcp_test.go            NEW — 140 lines — MCPTool/MCPResource route
                         tests + schema derivation + manifest inclusion.

Reviewer-flagged small fixes:

  examples/template/storage.go — dropped dead `_ = context.Background()`
    line + unused context import. Was a placeholder that landed in the
    scaffold.

  examples/template/mcp.go — replaced hardcoded \"v0.1.0\" in the
    StatusResource example with a commented `moduleVersion` constant set
    to \"0.0.0-unknown\". Prevents a copy-pasted template from silently
    reporting the wrong version; the comment points at -ldflags as the
    canonical wiring.

  mcp.go — one-sentence addition to MCPTool godoc noting that generics
    provide compile-time signature safety but schemas are derived via
    reflection at registration and validated against incoming JSON at
    call time, NOT statically against the wire format. Prevents readers
    from assuming compile-time schema enforcement.

Skipped findings:

  - Resolve[T] silent-stub trap: explicit design decision from prior
    discussion (API shape locked in, real resolution ships later).
    Godoc already calls out the stub behavior.
  - internal/registry types in public ManifestPayload signatures:
    pre-existing convention across Routes/Schedules/Tasks/Permissions
    fields. Not introduced by this branch; fixing would require a
    separate scope.
  - Dedup pattern extraction in registry.go: each \"first-wins\" loop
    has different key fields and update semantics. Generic extraction
    would require a key-extractor callback and not actually reduce
    complexity.
  - Trimming verbose unexported-method godoc: has historical context
    (prod vs dev split, caching rationale) that would be lost.
  - assertPanics consolidation across packages: structural — test files
    in different packages can't share helpers without a testutil
    package. Not worth it for current call sites.

go test ./... + go vet + gofmt clean. Public API unchanged.
