+++ examples/template/cache.go
@@ -6,8 +6,8 @@ package main
 import (
 	"net/http"
 
-	ms "github.com/mirrorstack-ai/app-module-sdk"
 	"github.com/go-chi/chi/v5"
+	ms "github.com/mirrorstack-ai/app-module-sdk"
 )
 
 func init() {
