+++ examples/template/meter.go
@@ -10,8 +10,8 @@ import (
 	"log"
 	"net/http"
 
-	ms "github.com/mirrorstack-ai/app-module-sdk"
 	"github.com/go-chi/chi/v5"
+	ms "github.com/mirrorstack-ai/app-module-sdk"
 )
 
 func init() {
