+++ examples/template/storage.go
@@ -4,11 +4,10 @@ package main
 // Remove this file if the module doesn't upload/download files.
 
 import (
-	"context"
 	"net/http"
 
-	ms "github.com/mirrorstack-ai/app-module-sdk"
 	"github.com/go-chi/chi/v5"
+	ms "github.com/mirrorstack-ai/app-module-sdk"
 )
 
 func init() {
@@ -28,6 +27,5 @@ func initUpload(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	// Use storer to issue a presigned multipart upload URL, list objects, etc.
-	_ = context.Background()
 	w.WriteHeader(http.StatusNotImplemented)
 }
