+++ mirrorstack_test.go
@@ -707,346 +707,3 @@ func TestModule_ModuleSchema(t *testing.T) {
 		})
 	}
 }
-
-// ---- Describe / DependsOn / Needs / Resolve ----
-
-func TestDescribe_PopulatesRegistry(t *testing.T) {
-	t.Parallel()
-
-	m, _ := New(Config{ID: "demo"})
-	m.Describe("A demo module")
-	if got := m.registry.Description(); got != "A demo module" {
-		t.Errorf("Description = %q, want %q", got, "A demo module")
-	}
-}
-
-func TestDependsOn_AlwaysRequired(t *testing.T) {
-	t.Parallel()
-
-	m, _ := New(Config{ID: "demo"})
-	m.DependsOn("oauth-core")
-	deps := m.registry.Dependencies()
-	if len(deps) != 1 {
-		t.Fatalf("len(Dependencies) = %d, want 1", len(deps))
-	}
-	if deps[0].ID != "oauth-core" || deps[0].Optional {
-		t.Errorf("Dependencies[0] = %+v, want {ID:oauth-core, Optional:false}", deps[0])
-	}
-}
-
-func TestNeeds_DeclaresOptionalDep(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	handler := func(w http.ResponseWriter, r *http.Request) {}
-	wrapped := Needs("video", handler)
-	if wrapped == nil {
-		t.Fatal("Needs returned nil handler")
-	}
-
-	deps := m.registry.Dependencies()
-	if len(deps) != 1 {
-		t.Fatalf("len(Dependencies) = %d, want 1", len(deps))
-	}
-	if deps[0].ID != "video" || !deps[0].Optional {
-		t.Errorf("Dependencies[0] = %+v, want {ID:video, Optional:true}", deps[0])
-	}
-}
-
-func TestNeeds_HandlerPassesThrough(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	called := false
-	handler := func(w http.ResponseWriter, r *http.Request) { called = true }
-	wrapped := Needs("video", handler)
-
-	rec := httptest.NewRecorder()
-	wrapped(rec, httptest.NewRequest("GET", "/x", nil))
-	if !called {
-		t.Error("Needs wrapper did not call the underlying handler")
-	}
-}
-
-func TestNeeds_RequiredWinsOverNeedsForSameID(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	// Required first, then Needs with same id.
-	DependsOn("video")
-	_ = Needs("video", func(w http.ResponseWriter, r *http.Request) {})
-
-	deps := m.registry.Dependencies()
-	if len(deps) != 1 {
-		t.Fatalf("len(Dependencies) = %d, want 1 (dedup)", len(deps))
-	}
-	if deps[0].Optional {
-		t.Errorf("Dependencies[0].Optional = true, want false (required wins)")
-	}
-}
-
-func TestNeeds_NestedDeclaresMultipleDeps(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	_ = Needs("video", Needs("audit-log", func(w http.ResponseWriter, r *http.Request) {}))
-
-	deps := m.registry.Dependencies()
-	if len(deps) != 2 {
-		t.Fatalf("len(Dependencies) = %d, want 2", len(deps))
-	}
-	ids := map[string]bool{deps[0].ID: deps[0].Optional, deps[1].ID: deps[1].Optional}
-	for id, optional := range ids {
-		if !optional {
-			t.Errorf("dep %q: Optional = false, want true", id)
-		}
-	}
-	if !ids["video"] || !ids["audit-log"] {
-		t.Errorf("deps = %+v, want both video and audit-log", deps)
-	}
-}
-
-func TestResolve_UnregisteredReturnsZeroAndFalse(t *testing.T) {
-	t.Parallel()
-
-	type fakeClient struct{ X int }
-	got, ok := Resolve[fakeClient]("not-registered")
-	if ok {
-		t.Error("Resolve for unregistered id returned ok=true, want false")
-	}
-	if got.X != 0 {
-		t.Errorf("Resolve returned non-zero value %+v, want zero", got)
-	}
-}
-
-// ---- MCP integration ----
-
-type greetArgs struct {
-	Name string `json:"name"`
-}
-
-type greetResult struct {
-	Message string `json:"message"`
-}
-
-func TestMCPTool_RegistersAndInvokesViaRoute(t *testing.T) {
-	// No t.Parallel: newTestModuleWithSecret calls t.Setenv, which is incompatible with parallel.
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	MCPTool("greet", "Say hi", func(ctx context.Context, a greetArgs) (greetResult, error) {
-		return greetResult{Message: "hi " + a.Name}, nil
-	})
-
-	// Call via the mounted route.
-	body := strings.NewReader(`{"name":"greet","args":{"name":"world"}}`)
-	req := httptest.NewRequest("POST", "/__mirrorstack/mcp/tools/call", body)
-	req.Header.Set("X-MS-Internal-Secret", "secret")
-	rec := httptest.NewRecorder()
-	m.Router().ServeHTTP(rec, req)
-
-	if rec.Code != 200 {
-		t.Fatalf("tools/call status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	var resp struct {
-		Result greetResult `json:"result"`
-	}
-	if err := json.Unmarshal(rec.Body.Bytes(), &resp); err != nil {
-		t.Fatalf("decode: %v", err)
-	}
-	if resp.Result.Message != "hi world" {
-		t.Errorf("result.message = %q, want %q", resp.Result.Message, "hi world")
-	}
-}
-
-func TestMCPTool_ToolsListRequiresInternalAuth(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	MCPTool("greet", "Say hi", func(ctx context.Context, a greetArgs) (greetResult, error) {
-		return greetResult{Message: "hi"}, nil
-	})
-
-	// No secret header.
-	req := httptest.NewRequest("GET", "/__mirrorstack/mcp/tools/list", nil)
-	rec := httptest.NewRecorder()
-	m.Router().ServeHTTP(rec, req)
-	if rec.Code == 200 {
-		t.Errorf("tools/list without auth returned 200, want 401/403")
-	}
-}
-
-func TestMCPTool_SchemasDerivedFromStructs(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	MCPTool("greet", "Say hi", func(ctx context.Context, a greetArgs) (greetResult, error) {
-		return greetResult{}, nil
-	})
-
-	tool, ok := m.registry.MCPTool("greet")
-	if !ok {
-		t.Fatal("tool not registered")
-	}
-	// The JSON Schema must describe the args struct's fields.
-	if !strings.Contains(string(tool.InputSchema), `"name"`) {
-		t.Errorf("InputSchema missing 'name' field: %s", tool.InputSchema)
-	}
-	if !strings.Contains(string(tool.OutputSchema), `"message"`) {
-		t.Errorf("OutputSchema missing 'message' field: %s", tool.OutputSchema)
-	}
-}
-
-func TestMCPResource_RegistersAndReadsViaRoute(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	type statusOut struct {
-		Healthy bool `json:"healthy"`
-	}
-	MCPResource("status", "Module status", func(ctx context.Context) (statusOut, error) {
-		return statusOut{Healthy: true}, nil
-	})
-
-	req := httptest.NewRequest("GET", "/__mirrorstack/mcp/resources/read?name=status", nil)
-	req.Header.Set("X-MS-Internal-Secret", "secret")
-	rec := httptest.NewRecorder()
-	m.Router().ServeHTTP(rec, req)
-
-	if rec.Code != 200 {
-		t.Fatalf("resources/read status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	if !strings.Contains(rec.Body.String(), `"healthy":true`) {
-		t.Errorf("body = %s, want content with healthy:true", rec.Body.String())
-	}
-}
-
-func TestMCPTool_ManifestIncludesMCPSurface(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	MCPTool("greet", "Say hi", func(ctx context.Context, a greetArgs) (greetResult, error) {
-		return greetResult{}, nil
-	})
-
-	req := httptest.NewRequest("GET", "/__mirrorstack/platform/manifest", nil)
-	req.Header.Set("X-MS-Internal-Secret", "secret")
-	rec := httptest.NewRecorder()
-	m.Router().ServeHTTP(rec, req)
-
-	var payload system.ManifestPayload
-	if err := json.Unmarshal(rec.Body.Bytes(), &payload); err != nil {
-		t.Fatalf("decode: %v", err)
-	}
-	if len(payload.MCP.Tools) != 1 || payload.MCP.Tools[0].Name != "greet" {
-		t.Errorf("manifest.mcp.tools = %+v, want [greet]", payload.MCP.Tools)
-	}
-}
-
-func TestDependsOn_VersionConstraintStoredInManifest(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	DependsOn("oauth-core@^1.2.0")
-
-	deps := m.registry.Dependencies()
-	if len(deps) != 1 {
-		t.Fatalf("len(deps) = %d, want 1", len(deps))
-	}
-	if deps[0].ID != "oauth-core" {
-		t.Errorf("deps[0].ID = %q, want oauth-core", deps[0].ID)
-	}
-	if deps[0].Version != "^1.2.0" {
-		t.Errorf("deps[0].Version = %q, want ^1.2.0", deps[0].Version)
-	}
-	if deps[0].Optional {
-		t.Errorf("deps[0].Optional = true, want false")
-	}
-}
-
-func TestNeeds_VersionConstraint(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	_ = Needs("video@~1.2.0", func(w http.ResponseWriter, r *http.Request) {})
-
-	deps := m.registry.Dependencies()
-	if len(deps) != 1 {
-		t.Fatalf("len(deps) = %d, want 1", len(deps))
-	}
-	if deps[0].ID != "video" || deps[0].Version != "~1.2.0" || !deps[0].Optional {
-		t.Errorf("deps[0] = %+v, want {video, ~1.2.0, optional}", deps[0])
-	}
-}
-
-func TestDependsOn_NoConstraintLeavesVersionEmpty(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	DependsOn("oauth-core")
-
-	deps := m.registry.Dependencies()
-	if len(deps) != 1 || deps[0].Version != "" {
-		t.Errorf("deps[0].Version = %q, want empty", deps[0].Version)
-	}
-}
-
-func TestDependsOn_InvalidConstraintPanics(t *testing.T) {
-	resetDefault(t)
-	m := newTestModuleWithSecret(t, "demo")
-	defaultModule = m
-
-	defer func() {
-		if r := recover(); r == nil {
-			t.Error("expected panic on invalid SemVer constraint")
-		}
-	}()
-	DependsOn("oauth-core@not-a-semver")
-}
-
-func TestParseDepSpec_SupportsCommonFormats(t *testing.T) {
-	t.Parallel()
-
-	cases := []struct {
-		spec       string
-		wantID     string
-		wantVer    string
-		wantPanic  bool
-	}{
-		{"oauth-core", "oauth-core", "", false},
-		{"oauth-core@^1.2.0", "oauth-core", "^1.2.0", false},
-		{"oauth-core@~1.2.0", "oauth-core", "~1.2.0", false},
-		{"oauth-core@1.2.3", "oauth-core", "1.2.3", false},
-		{"oauth-core@>=1.2.0", "oauth-core", ">=1.2.0", false},
-		{"oauth-core@1.x", "oauth-core", "1.x", false},
-		{"oauth-core@", "oauth-core", "", false}, // empty constraint = any
-	}
-	for _, tc := range cases {
-		t.Run(tc.spec, func(t *testing.T) {
-			defer func() {
-				r := recover()
-				if tc.wantPanic && r == nil {
-					t.Errorf("expected panic for spec=%q", tc.spec)
-				} else if !tc.wantPanic && r != nil {
-					t.Errorf("unexpected panic for spec=%q: %v", tc.spec, r)
-				}
-			}()
-			gotID, gotVer := parseDepSpec(tc.spec)
-			if gotID != tc.wantID || gotVer != tc.wantVer {
-				t.Errorf("parseDepSpec(%q) = (%q, %q), want (%q, %q)", tc.spec, gotID, gotVer, tc.wantID, tc.wantVer)
-			}
-		})
-	}
-}
